BL  for price and product; Product Window added.

+++ DesktopEnvironmentManager/MainWindow.xaml.cs
@@ -33,8 +33,8 @@ namespace DesktopEnvironmentManager
         private void Button_Click(object sender, RoutedEventArgs e)
         {
             // getting the values 
-            var productTextBox = int.Parse(ProductText.Text);
-            var priceTextBox = int.Parse(PriceText.Text);
+            var productTextBox = double.Parse(ProductText.Text);
+            var priceTextBox = double.Parse(PriceText.Text);
 
             //show the value 
             // MessageBox.Show("This is your total: " + (priceTextBox * productTextBox));
